Refine `utils.py` usage Define path variables in utils, move single use functions out of utils, and add docstrings to all the tests
